+++ crates/web-sys/src/features/gen_PublicKeyCredentialRpEntity.rs
@@ -13,13 +13,13 @@ extern "C" {
     #[doc = "Get the `icon` field of this object."]
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `PublicKeyCredentialRpEntity`*"]
-    #[deprecated]
+    #[deprecated(note = "Removed from the WebAuthn specification")]
     #[wasm_bindgen(method, getter = "icon")]
     pub fn get_icon(this: &PublicKeyCredentialRpEntity) -> Option<::alloc::string::String>;
     #[doc = "Change the `icon` field of this object."]
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `PublicKeyCredentialRpEntity`*"]
-    #[deprecated]
+    #[deprecated(note = "Removed from the WebAuthn specification")]
     #[wasm_bindgen(method, setter = "icon")]
     pub fn set_icon(this: &PublicKeyCredentialRpEntity, val: &str);
     #[doc = "Get the `name` field of this object."]
@@ -53,7 +53,7 @@ impl PublicKeyCredentialRpEntity {
         ret.set_name(name);
         ret
     }
-    #[deprecated = "Use `set_icon()` instead."]
+    #[deprecated(note = "Removed from the WebAuthn specification")]
     pub fn icon(&mut self, val: &str) -> &mut Self {
         self.set_icon(val);
         self